requirements_antigo.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.1
dataclasses-json==0.6.7
decorator==5.2.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@536dc9d527074e3b15df5f6677ffe1f4e104a4ab
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
git-filter-repo==2.47.0
GPUtil==1.4.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
imgviz==1.7.6
iopath==0.1.9
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy_loader==0.4
lmdb==1.6.2
loguru==0.7.3
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
natsort==8.4.0
networkx==3.4.2
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.63.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
osam==0.2.3
packaging==24.2
paddle-bfloat==0.1.7
pandas==1.5.3
pathspec==0.12.1
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==11.2.1
platformdirs==4.3.8
portalocker==3.1.1
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.1
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.20.2
pyparsing==3.2.3
pypdfium2==4.30.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.19.1
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.133
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.1.3
win32_setctime==1.2.0
yacs==0.1.8
yarl==1.20.0
